Create colorama.php
Using a HTML, create a table of 3 rows by 3 columns such that each td is considered a box when viewed on the browser. Randomly set a background color to each of the 9 boxes. These should be varying colors which can only be any of red, green or yellow. Use PHP to randomly set these background colors. To do this, use the PHP rand() function to generate random values between 1 to 3 such that each of the 3 colors are represented by a number. Call the PHP code in the background-color property.